fix(ci): satisfy source file line limit
