+++ app/lib/server/applications/decision.ts
@@ -1,10 +1,9 @@
 "use server";
 
-import { APPLICATION_STATUS_LABELS } from "../../applications/status";
 import { assertApplicationAdmissionReady } from "../../applications/admissionEligibility";
 import type { ApplicationDecision } from "../../applications/decisionEmail";
 import { isApplicationStatusTransitionAllowed } from "../../applications/transitions";
-import { applications, jobPostings } from "../../db/collections";
+import { jobPostings } from "../../db/collections";
 import type { Application } from "../../db/types";
 import { sendUserStateEmail } from "../users/email";
 import { loadOwnedApplication } from "./access";
@@ -13,9 +12,8 @@ import { prepareAcceptance, sendDecisionEmail } from "./decisionDelivery";
 import {
   applicationDecisionInputSchema,
   type ApplicationDecisionInput,
-  type ParsedApplicationDecision,
 } from "./decisionInput";
-import { createApplicationHistoryEntry } from "./history";
+import { persistDecision } from "./decisionPersistence";
 import {
   createAcceptedApplicantMember,
   rollbackAcceptedApplicantMember,
@@ -136,77 +134,4 @@ function assertDecisionAllowed(
   if (decision === "rejected" && isProvisioning) {
     throw new Error("Das Workspace-Konto wird bereits eingerichtet");
   }
-}
-
-async function persistDecision(input: {
-  application: Application;
-  decision: ParsedApplicationDecision;
-  organizationId: string;
-  userId: string;
-  workspaceUserId?: string;
-  onboardingUserId?: string;
-}): Promise<string> {
-  const entry = createApplicationHistoryEntry(
-    input.userId,
-    "status_changed",
-    `${APPLICATION_STATUS_LABELS[input.application.status]} → ${APPLICATION_STATUS_LABELS[input.decision.decision]}`,
-    {
-      fromStatus: input.application.status,
-      toStatus: input.decision.decision,
-    },
-  );
-  const result = await (
-    await applications()
-  ).updateOne(
-    {
-      _id: input.application._id,
-      organizationId: input.organizationId,
-      status: input.application.status,
-      ...(input.workspaceUserId
-        ? { workspaceProvisioningStatus: "provisioned" }
-        : {}),
-      ...(input.onboardingUserId
-        ? {
-            $or: [
-              { onboardingUserId: { $exists: false } },
-              { onboardingUserId: input.onboardingUserId },
-            ],
-          }
-        : {}),
-    },
-    {
-      $set: {
-        status: input.decision.decision,
-        updatedAt: entry.timestamp,
-        ...(input.workspaceUserId
-          ? { workspaceProvisioningStatus: "invited" as const }
-          : {}),
-        ...(input.onboardingUserId
-          ? {
-              onboardingUserId: input.onboardingUserId,
-              onboardingLinkedAt: entry.timestamp,
-              onboardingStartedAt: entry.timestamp,
-              onboardingStartedBy: input.userId,
-              cleanupEligibleAt: entry.timestamp,
-            }
-          : {}),
-      },
-      ...(input.workspaceUserId || input.onboardingUserId
-        ? {
-            $unset: {
-              ...(input.workspaceUserId
-                ? { workspaceProvisioningError: "" }
-                : {}),
-              ...(input.onboardingUserId ? { onboardingLinkError: "" } : {}),
-            },
-          }
-        : {}),
-      $push: { history: entry },
-    },
-  );
-  if (result.modifiedCount !== 1) {
-    throw new Error("Bewerbung wurde zwischenzeitlich geändert");
-  }
-
-  return entry.details;
 }
